+++ lib/Controller/PaymentMethod/OfflineController.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace IA\PaymentBundle\Controller\PaymentMethod;
+namespace Vankosoft\PaymentBundle\Controller\PaymentMethod;
 
 use Payum\Bundle\PayumBundle\Controller\PayumController;
 
-class PaymentController extends PayumController
+class OfflineController extends PayumController
 {
     
     /**