+++ lib/DependencyInjection/Configuration.php
@@ -2,6 +2,24 @@
 
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
+use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
+
+use Sylius\Bundle\ResourceBundle\SyliusResourceBundle;
+use Sylius\Bundle\ResourceBundle\Doctrine\ORM\EntityRepository;
+use Sylius\Component\Resource\Factory\Factory;
+
+use Vankosoft\PaymentBundle\Model\GatewayConfig;
+use Vankosoft\PaymentBundle\Model\Interfaces\GatewayConfigInterface;
+use Vankosoft\PaymentBundle\Controller\GatewayConfigController;
+use Vankosoft\PaymentBundle\Form\GatewayConfigForm;
+
+use Vankosoft\PaymentBundle\Model\PaymentMethod;
+use Vankosoft\PaymentBundle\Model\Interfaces\PaymentMethodInterface;
+use Vankosoft\PaymentBundle\Controller\PaymentMethodController;
+use Vankosoft\PaymentBundle\Form\PaymentMethodForm;
+
+use Vankosoft\PaymentBundle\Model\Payment;
+use Vankosoft\PaymentBundle\Model\Interfaces\PaymentInterface;
 
 /**
  * This is the class that validates and merges configuration from your app/config files
@@ -15,21 +33,89 @@ class Configuration implements ConfigurationInterface
      */
     public function getConfigTreeBuilder()
     {
-        $treeBuilder = new TreeBuilder();
-//         $rootNode = $treeBuilder->root('ia_payment');
-
-//         $rootNode->children()
-//             ->arrayNode('payment_accounts')->isRequired()
-//             ->prototype('variable')
-//             ->treatNullLike(array())
-//         ;
+        $treeBuilder    = new TreeBuilder( 'vs_payment' );
+        $rootNode       = $treeBuilder->getRootNode();
         
-//         $rootNode->children()
-//             ->arrayNode('payment_methods')->isRequired()
-//                 ->prototype('variable')
-//                 ->treatNullLike(array())
-//         ;
+        $rootNode
+            ->children()
+                ->scalarNode( 'orm_driver' )
+                    ->defaultValue( SyliusResourceBundle::DRIVER_DOCTRINE_ORM )->cannotBeEmpty()
+                ->end()
+        //             ->arrayNode('payment_accounts')->isRequired()
+        //             ->prototype('variable')
+        //             ->treatNullLike(array())
+        //         ;
+        
+        //         $rootNode->children()
+        //             ->arrayNode('payment_methods')->isRequired()
+        //                 ->prototype('variable')
+        //                 ->treatNullLike(array())
+            ->end()
+        ;
+        $this->addResourcesSection( $rootNode );
 
         return $treeBuilder;
+    }
+    
+    private function addResourcesSection( ArrayNodeDefinition $node ): void
+    {
+        $node
+            ->children()
+                ->arrayNode( 'resources' )
+                    ->addDefaultsIfNotSet()
+                    ->children()
+                        ->arrayNode( 'gateway_config' )
+                            ->addDefaultsIfNotSet()
+                            ->children()
+                                ->variableNode( 'options' )->end()
+                                ->arrayNode( 'classes' )
+                                    ->addDefaultsIfNotSet()
+                                    ->children()
+                                        ->scalarNode( 'model' )->defaultValue( GatewayConfig::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'interface' )->defaultValue( GatewayConfigInterface::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'controller' )->defaultValue( GatewayConfigController::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'repository' )->defaultValue( EntityRepository::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'factory' )->defaultValue( Factory::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'form' )->defaultValue( GatewayConfigForm::class )->cannotBeEmpty()->end()
+                                    ->end()
+                                ->end()
+                            ->end()
+                        ->end()
+                        ->arrayNode( 'payment_method' )
+                            ->addDefaultsIfNotSet()
+                            ->children()
+                                ->variableNode( 'options' )->end()
+                                ->arrayNode( 'classes' )
+                                    ->addDefaultsIfNotSet()
+                                    ->children()
+                                        ->scalarNode( 'model' )->defaultValue( PaymentMethod::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'interface' )->defaultValue( PaymentMethodInterface::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'controller' )->defaultValue( PaymentMethodController::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'repository' )->defaultValue( EntityRepository::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'factory' )->defaultValue( Factory::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'form' )->defaultValue( PaymentMethodForm::class )->cannotBeEmpty()->end()
+                                    ->end()
+                                ->end()
+                            ->end()
+                        ->end()
+                        ->arrayNode( 'payment' )
+                            ->addDefaultsIfNotSet()
+                            ->children()
+                                ->variableNode( 'options' )->end()
+                                ->arrayNode( 'classes' )
+                                    ->addDefaultsIfNotSet()
+                                    ->children()
+                                        ->scalarNode( 'model' )->defaultValue( Payment::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'interface' )->defaultValue( PaymentInterface::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'repository' )->defaultValue( EntityRepository::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'factory' )->defaultValue( Factory::class )->cannotBeEmpty()->end()
+                                    ->end()
+                                ->end()
+                            ->end()
+                        ->end()
+                    ->end()
+                ->end()
+            ->end()
+        ;
     }
 }