+++ lib/Form/PaymentMethodForm.php
@@ -1,6 +1,6 @@
-<?php namespace IA\PaymentBundle\Form;
+<?php namespace Vankosoft\PaymentBundle\Form;
 
-use Symfony\Component\Form\AbstractType;
+use Vankosoft\ApplicationBundle\Form\AbstractForm;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
@@ -10,20 +10,14 @@ use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\Extension\Core\Type\ButtonType;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 
-use IA\PaymentBundle\Entity\PaymentMethod as PaymentMethodData;
-use IA\PaymentBundle\Entity\GatewayConfig;
+use Vankosoft\PaymentBundle\Entity\PaymentMethod as PaymentMethodData;
+use Vankosoft\PaymentBundle\Entity\GatewayConfig;
 
 /**
  * Credit Card Form Type for PayPal Pro Direct Payments
  */
-class PaymentMethod extends AbstractType
+class PaymentMethodForm extends AbstractForm
 {
-
-    public function getName()
-    {
-        return 'ia_payment_paymentmethod';
-    }
-
     public function buildForm( FormBuilderInterface $builder, array $options )
     {
         $builder
@@ -35,21 +29,18 @@ class PaymentMethod extends AbstractType
             ->add( 'name', TextType::class )
             ->add( 'route', TextType::class )
             ->add( 'active', CheckboxType::class, array('required'=>false ) )
-                
-            
-            
-            ->add('btnSave', SubmitType::class, array('label' => 'Save'))
-            ->add('btnCancel', ButtonType::class, array('label' => 'Cancel'))
         ;
     }
 
-    public function configureOptions(OptionsResolver $resolver)
+    public function configureOptions( OptionsResolver $resolver ): void
     {
-        $resolver->setDefaults(array(
-            'data_class' => PaymentMethodData::class
-        ));
+        parent::configureOptions( $resolver );
     }
 
+    public function getName()
+    {
+        return 'vs_payment.payment_method';
+    }
 }