+++ lib/Model/Interfaces/PaymentInterface.php
@@ -1,0 +1,8 @@
+<?php namespace Vankosoft\PaymentBundle\Model\Interfaces;
+
+use Sylius\Component\Resource\Model\ResourceInterface;
+
+interface PaymentInterface extends ResourceInterface
+{
+    
+}