+++ lib/Model/Token.php
@@ -1,12 +1,12 @@
 <?php
-namespace IA\PaymentBundle\Entity;
+namespace Vankosoft\PaymentBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
 use Payum\Core\Model\Token as BaseToken;
 
 /**
- * @ORM\Table(name="IAP_Tokens")
- * @ORM\Entity
+ * ORM\Table(name="IAP_Tokens")
+ * ORM\Entity
  */
 class Token extends BaseToken
 {