+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+0.0.2	|	Release date: **12.05.2022**
+============================================
+* Bug-Fixes, Refactoring and Improvements:
+  - Add Default ORM Driver in Configuration.
+  - Big Refactoring.
+  - Fix Routes.
+  - Refactoring and Fixing Templates and Payment Configs AdminPanel Menu.
+  - Fixes for Gateway Config Form Template.
+  - Fixes for Gateway Config Form Template.
+  - Add All Translations for Gateway Config Form and Templates.
+  - Add All Translations for Gateway Config Form and Templates.
+  - Fix GatewayConfig Options Form.
+  - Add Defaults For Route Parameter to Can Be Added in Menu.
+  - Fix GatewayConfigExtController.
+  - Prepare Payment Routes.
+  - Improve composer.json
+
+
 0.0.1	|	Release date: **12.03.2022**
 ============================================
 * New Features: